+++ src/main/java/structures/Person.java
@@ -1,4 +1,4 @@
-package tests.lesson04;
+package structures;
 
 public class Person {
