+++ src/main/resources/properties/settings.properties
@@ -1,1 +1,3 @@
 webDriver=chrome
+timeout=5
+baseUrl=http://blazedemo.com
